requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
amqp==5.3.1
    # via kombu
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
billiard==4.2.1
    # via celery
celery==5.5.3
    # via -r requirements.in
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django==5.2.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-phonenumber-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==4.7.0
    # via -r requirements.in
django-debug-toolbar==6.0.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-phonenumber-field==8.1.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-yasg==1.21.10
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-yasg
kombu==5.5.4
    # via celery
packaging==25.0
    # via
    #   drf-yasg
    #   gunicorn
    #   kombu
phonenumbers==9.0.13
    # via -r requirements.in
pillow==11.3.0
    # via -r requirements.in
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.2
    # via drf-yasg
redis==6.4.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
tzdata==2025.2
    # via kombu
uritemplate==4.2.0
    # via drf-yasg
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.9.0
    # via -r requirements.in
